requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
gunicorn==23.0.0
jinja2==3.1.6
sqlalchemy==2.0.43
alembic==1.16.5
python-jose[cryptography]==3.5.0
pwdlib==0.2.1
argon2-cffi==25.1.0
python-multipart==0.0.20
pydantic-settings==2.10.1
reportlab==4.4.4
pillow==11.3.0
psycopg[binary]==3.2.13
slowapi==0.1.9
httpx==0.28.1
translators>=5.9.2
pytest==8.4.2
pytest-asyncio==1.2.0
playwright==1.54.0
beautifulsoup4==4.13.4
